Update
Added instruction list, so now assemble() can just reference that, and
the functions themselves will fill in the bytes and such, but there will
be helper functions for that.

I added in a label struct, this struct will basically house information
about labels. I now need to think of some sort of system where local
labels would be functional.

I added some more characters and logic to the lex().

I also made a simple enum to control where each instruction is placed in
the instruction list and the ISA.