Instructions
I have lex'd in instructions, and have made 4 functions for
the parsing of instructions. These are generic functions, for a two
argument instruction, one, and zero, as well
as a function for an instruction with an unknown number of arguments.
This should work as I am able to specify syntax with
a series of functions within these generic functions.
For example a instruction with two arguments should have syntax:
instruction
size (maybe)
argument
comma
size (maybe)
argument