Issue fixed
Bad logic, a new string was being grabbed for every comparsion.
I just made it so that the string would be grabbed once, the functions
had access to a copy of it where it is uppercase.
Then the functions themselves would make their own strings to apply
their own specific filters.