Implemented carry over
The conditions of the while loop were still too simple to be effective
at ensuring that lexing would be done line by line. I've implemented a
carry over so that if the assembler reads too far ahead, the carry over
can be set as the head, and have the token which is on the next line
be the first token of the next. The current method does cause a
segmentation fault.