Fixed
I was just memcpying the data of the carry over token into the current
token, rather than actually setting the data of the current token
to the carry over's. Thus when the carry over was reset, the extra_bytes
member was also freed, lost, and caused a segmentation fault.
The carry over method still doesn't work for test4.asm