Simple processing
Implemented the basic processing of the bytes.
I made it so that there is a definition which will do the operation
and then a function which tell the definition what variables to
manipulate, this is what I was trying to implement in the original
emulator, but it didn't work out.