Interrupts and lex
I added in interrupts, but I discovered the my assembler
did not suppoer hex values, and or doing an ascii character
in code. I quickly implemented that into the assembler so that it would
recognize hex and chars. I have also discovered that I am unable to
define strings or define bytes, which I should also implement.