Added wrh stack
I made it so that the current position that the assembler
is writing, or the write head's position, can be saved and restored
allowing for even wackier manipulation of the output code.
I also replaced the hash checking with a simple for loop in the lexer
and a couple of arrays.