New instructions
I added more varients to the mov instruction, so that if a
register were to be the first operand, then it would have
its own opcode. This way it is easier to translate into micro
code. For this reason, I changed back the order of the operands
to be in order as they are read.
I changed them originally to make move
register instructions easier, but with new instructions the reversion
has been reversed as it is no longer required.
The instruction can also be done faster.