Compression
The reference compressor does not work with addresses
which are more than 1 byte in length, byte it tries its best.
I also tested test11.asm with nasm, of course porting it.
However, I found that NASM is 5 bytes more efficient than my assembler,
which in my opinion is pretty cool.