Working out some stuff
I came up with a new encoding, Szikra V3.

For this, I've made the assembler and the emulator compatible with it.

So far I've made it that the assembler can write both v2
and v3 encodings into a given output file.

I've separated the v2 instruction processing into its
own file, and I am now going to be working on the v3
one. The v, or varient, can be specified in both assembler and emulator
using the -v tag.

I used a lot of if statements in the assembler, so I would like to clean
that up by separating the v2 and v3 encodings into their own files.